chore(sdk): trim local build/link overhead and add CI sccache (spectrum-design-data-tdb)
Investigated the reported 4+ minute cargo test/build times and a 23m+
"hanging" test. The test itself is not hanging (depth-capped recursion,
microseconds of work) and the root cause is not compile or link time either
— both are fast (15-35s) once warm. The actual bottleneck is macOS
Gatekeeper/XProtect on-exec scanning of freshly-built test binaries, an
MDM-enforced OS policy outside this repo's control (full writeup in the
bead notes).

- sdk/Cargo.toml: line-tables-only debuginfo + unpacked split-debuginfo for
  dev/test profiles — a low-risk trim, though not the dominant cost.
- sdk/core/tests/{prop_naming,suggest_calibration}.rs merged into
  sdk/core/tests/integration.rs — fewer compiled test binaries means fewer
  first-exec Gatekeeper/XProtect scans per `cargo test` run.
- sdk/scripts/bench-build.sh: reusable incremental-build benchmark used to
  measure this investigation.
- .github/workflows/ci.yml: sccache wired into the rust job's cold
  ~455-crate compile, layered with the existing rust-cache.
